split for seperate config file, and save config to logs for every training
