+++ samples/braintissue/braintissue.py
@@ -43,7 +43,10 @@ import datetime
 import time
 import numpy as np
 import skimage.io
+import pickle
 from imgaug import augmenters as iaa
+from shutil import copyfile
+import re
 
 # Root directory of the project
 ROOT_DIR = os.path.abspath("../../")
@@ -55,108 +58,7 @@ from mrcnn import utils
 from mrcnn import model as modellib
 from mrcnn import visualize
 
-# Path to trained weights file
-COCO_WEIGHTS_PATH = os.path.join(ROOT_DIR, "weights/mask_rcnn_coco.h5")
-
-# Directory to save logs and model checkpoints, if not provided
-# through the command line argument --logs
-DEFAULT_LOGS_DIR = os.path.join(ROOT_DIR, "logs")
-
-# Results directory
-# Save submission files here
-RESULTS_DIR = os.path.join(ROOT_DIR, "results/braintissue/")
-
-# The dataset doesn't have a standard train/val split, so I picked
-# a variety of images to serve as a validation set.
-VAL_IMAGE_IDS = ["artif1_0_crop1", "artif1_6_crop4", "artif1_2_crop8"]
-
-############################################################
-#  Configurations
-############################################################
-
-class BraintissueConfig(Config):
-    """Configuration for training on the Braintissue segmentation dataset."""
-    # Give the configuration a recognizable name
-    NAME = "Braintissue"
-
-    # Adjust depending on your GPU memory
-    IMAGES_PER_GPU = 1
-
-    # Number of classes (including background)
-    NUM_CLASSES = 1 + 1 + 1 # Background + Braintissue + magnetic part
-
-    # Number of training and validation steps per epoch
-    STEPS_PER_EPOCH = (100 - len(VAL_IMAGE_IDS)) // IMAGES_PER_GPU
-    VALIDATION_STEPS = 1 # max(1, len(VAL_IMAGE_IDS) // IMAGES_PER_GPU)
-
-    # Backbone network architecture
-    # Supported values are: resnet50, resnet101
-    BACKBONE = "resnet50"
-
-    # Input image resizing
-    IMAGE_RESIZE_MODE = "square"
-    IMAGE_MIN_DIM = 512
-    IMAGE_MAX_DIM = 512
-    IMAGE_MIN_SCALE = 0
-
-    # Length of square anchor side in pixels
-    RPN_ANCHOR_SCALES = (32, 64, 128, 256) # (8, 16, 32, 64, 128)
-
-    # ROIs kept after non-maximum supression (training and inference)
-    POST_NMS_ROIS_TRAINING = 128
-    POST_NMS_ROIS_INFERENCE = 128
-
-    # Non-max suppression threshold to filter RPN proposals.
-    # You can increase this during training to generate more proposals.
-    RPN_NMS_THRESHOLD = 0.9
-
-    # How many anchors per image to use for RPN training
-    RPN_TRAIN_ANCHORS_PER_IMAGE = 64
-
-    # Grayscale adjustments
-    IMAGE_CHANNEL_COUNT = 1
-    MEAN_PIXEL = np.array([184])
-    
-
-    # If enabled, resizes instance masks to a smaller size to reduce
-    # memory load. Recommended when using high-resolution images.
-    USE_MINI_MASK = True
-    MINI_MASK_SHAPE = (56, 56)  # (height, width) of the mini-mask
-
-    # Number of ROIs per image to feed to classifier/mask heads
-    # The Mask RCNN paper uses 512 but often the RPN doesn't generate
-    # enough positive proposals to fill this and keep a positive:negative
-    # ratio of 1:3. You can increase the number of proposals by adjusting
-    # the RPN NMS threshold.
-    TRAIN_ROIS_PER_IMAGE = 128
-
-    # Maximum number of ground truth instances to use in one image
-    MAX_GT_INSTANCES = 1000	
-
-    # Max number of final detections per image
-    DETECTION_MAX_INSTANCES = 100
-
-    # Loss weights for more precise optimization.
-    # Can be used for R-CNN training setup.
-    LOSS_WEIGHTS = {
-        "rpn_class_loss": 1.,
-        "rpn_bbox_loss": 2.,
-        "mrcnn_class_loss": 1.,
-        "mrcnn_bbox_loss": 1.,
-        "mrcnn_mask_loss": 2.
-    }
-
-
-class BraintissueInferenceConfig(BraintissueConfig):
-    # Set batch size to 1 to run one image at a time
-    GPU_COUNT = 1
-    IMAGES_PER_GPU = 1
-    # Don't resize imager for inferencing
-    IMAGE_RESIZE_MODE = "square"
-    # Non-max suppression threshold to filter RPN proposals.
-    # You can increase this during training to generate more proposals.
-    RPN_NMS_THRESHOLD = 0.7
-
+from braintissue_config import *
 
 ############################################################
 #  Dataset
@@ -173,13 +75,13 @@ class BraintissueDataset(utils.Dataset):
                 * train: stage1_train excluding validation images
                 * val: validation images from VAL_IMAGE_IDS
         """
-        # Add classes. We have one class.
+        # Add classes. We have two classes: magnetic part and braintissue part
         # Naming the dataset Braintissue, and the class Braintissue
         self.add_class("Braintissue", 1, "Braintissue")
         self.add_class("Braintissue", 2, "Magnet")
 
         # Which subset?
-        # "val": use hard-coded list above [UNAVAILABLE -Niklas ]
+        # "val": use hard-coded list above
         # "train": use data from stage1_train minus the hard-coded list above
         # else: use the data from the specified sub-directory
         assert subset in ["train", "val", "stage1_train", "stage1_test", 
@@ -412,6 +314,28 @@ def detect(model, dataset_dir, subset):
         f.write(submission)
     print("Saved to ", submit_dir)
 
+############################################################
+#  Config saving
+############################################################
+
+def save_config(logs_path):
+    """ saves the braintissue config to the log dir, to have it next to the weights """
+
+    now = datetime.datetime.now()
+
+    # TODO: model.py 2335 possible conflict, if minute differs
+    log_dir = os.path.join(logs_path, 
+                            "{}{:%Y%m%dT%H%M}".format(config.NAME.lower(), now))
+
+    # Create log_dir if it does not exist
+    if not os.path.exists(log_dir):
+        os.makedirs(log_dir)
+
+    dest = os.path.join(log_dir, f"braintissue_config.py")
+
+    # Copy braintissue config to log dir
+    copyfile("braintissue_config.py", dest)
+
 
 ############################################################
 #  Command Line
@@ -422,7 +346,7 @@ if __name__ == '__main__':
 
     # Parse command line arguments
     parser = argparse.ArgumentParser(
-        description='Mask R-CNN for nuclei counting and segmentation')
+        description='Mask R-CNN for braintissue wafer segmentation')
     parser.add_argument("command",
                         metavar="<command>",
                         help="'train' or 'detect'")
@@ -456,6 +380,9 @@ if __name__ == '__main__':
     # Configurations
     if args.command == "train":
         config = BraintissueConfig()
+
+        # Copy the used config file to the log dir for reproducibility after training
+        save_config(args.logs)
     else:
         config = BraintissueInferenceConfig()
     config.display()
